YA: from usize to u32 in spimlib. usize is 8 bytes...

+++ tpx3/src/clusterlib.rs
@@ -274,7 +274,7 @@ pub mod cluster {
             }
         }
 
-        pub fn get_or_not_spim_index(&self, spim_tdc: Option<PeriodicTdcRef>, xspim: usize, yspim: usize) -> Option<usize> {
+        pub fn get_or_not_spim_index(&self, spim_tdc: Option<PeriodicTdcRef>, xspim: usize, yspim: usize) -> Option<u32> {
             if let Some(frame_tdc) = spim_tdc {
                 spimlib::get_spimindex(self.x(), self.frame_dt(), &frame_tdc, xspim, yspim)
             } else {
